+++ src/twg2/ast/interm/annotation/AnnotationSig.java
@@ -45,4 +45,48 @@ public class AnnotationSig implements JsonWritableSig {
 		return "annotation: " + name + "(" + arguments + ")";
 	}
 
+
+	@Override
+	public int hashCode() {
+		final int prime = 31;
+		int result = 1;
+		result = prime * result + (arguments == null ? 0 : arguments.hashCode());
+		result = prime * result + (fullName == null ? 0 : fullName.hashCode());
+		result = prime * result + (name == null ? 0 : name.hashCode());
+		return result;
+	}
+
+
+	@Override
+	public boolean equals(Object obj) {
+		if (this == obj)
+			return true;
+		if (obj == null || this.getClass() != obj.getClass())
+			return false;
+
+		AnnotationSig other = (AnnotationSig) obj;
+		if (arguments == null) {
+			if (other.arguments != null)
+				return false;
+		}
+		else if (!arguments.equals(other.arguments))
+			return false;
+
+		if (fullName == null) {
+			if (other.fullName != null)
+				return false;
+		}
+		else if (!fullName.equals(other.fullName))
+			return false;
+
+		if (name == null) {
+			if (other.name != null)
+				return false;
+		}
+		else if (!name.equals(other.name))
+			return false;
+
+		return true;
+	}
+
 }