+++ test/twg2/parser/test/utils/MethodAssert.java
@@ -1,10 +1,10 @@
 package twg2.parser.test.utils;
 
-import java.util.Arrays;
 import java.util.List;
 
 import org.junit.Assert;
 
+import twg2.ast.interm.annotation.AnnotationSig;
 import twg2.ast.interm.method.ParameterSig;
 import twg2.parser.codeParser.AccessModifier;
 
@@ -14,17 +14,15 @@ import twg2.parser.codeParser.AccessModifier;
  */
 public class MethodAssert {
 
-	public static void assertParameter(List<ParameterSig> params, int idx, String name, String type, List<? extends AccessModifier> accessMods, List<String> comments) {
+	public static void assertParameter(List<ParameterSig> params, int idx, String name, String type, List<? extends AccessModifier> accessMods, List<? extends AnnotationSig> annotations) {
 		ParameterSig p = params.get(idx);
 		Assert.assertEquals(name, p.getName());
 		Assert.assertEquals(type, p.getTypeSimpleName());
 		if(accessMods != null) {
 			Assert.assertEquals(accessMods, p.getParameterModifiers());
 		}
-		if(comments != null) {
-			Assert.assertEquals(Arrays.asList(" <summary>Add name</summary>\n",
-					" <param name=\"name\">the name</param>\n",
-					" <returns>the names</returns>\n"), comments);
+		if(annotations != null) {
+			Assert.assertEquals(annotations, p.getAnnotations());
 		}
 	}